Composer Trait in ServiceBubble
